+++ collect/tracehound-collect.sh
@@ -34,7 +34,7 @@
 
 set -eu
 
-VERSION="1.0.1"
+VERSION="1.0.2"
 OUT=""
 REFERENCE=""
 HOSTNAME_OVERRIDE=""